data_ingestion/requirements.txt
fastapi
uvicorn
google-cloud-bigquery
google-auth
pydantic
